+++ src/shopper-insights/shopperSession.test.js
@@ -32,8 +32,6 @@ const mockFindEligiblePaymentsRequest = (
 const defaultSdkConfig = {
   sdkToken: "sdk client token",
   pageType: "checkout",
-  clientToken: "",
-  userIDToken: "",
   paypalApiDomain: "https://api.paypal.com",
   environment: "test",
   buyerCountry: "US",
@@ -65,22 +63,23 @@ afterEach(() => {
   vi.clearAllMocks();
 });
 
-describe("shopper insights component - isEligibleInPaypalNetwork()", () => {
+describe("shopper insights component - isEligibleInPayPalNetwork()", () => {
   test("should get is member using the shopper insights API", async () => {
     const shopperSession = createShopperSession();
-    const recommendedPaymentMethods = await shopperSession.isEligibleInPaypalNetwork({
-      email: "email@test.com",
-      phone: {
-        countryCode: "1",
-        nationalNumber: "2345678901",
-      },
-    });
+    const recommendedPaymentMethods =
+      await shopperSession.isEligibleInPayPalNetwork({
+        email: "email@test.com",
+        phone: {
+          countryCode: "1",
+          nationalNumber: "2345678901",
+        },
+      });
 
     expect.assertions(1);
     expect(recommendedPaymentMethods).toEqual(true);
   });
 
-  test("should return isEligibleInPaypalNetwork true as long as one payment method is true", async () => {
+  test("should return isEligibleInPayPalNetwork true as long as one payment method is true", async () => {
     const shopperSession = createShopperSession({
       request: () =>
         Promise.resolve({
@@ -95,19 +94,20 @@ describe("shopper insights component - isEligibleInPaypalNetwork()", () => {
         }),
     });
 
-    const recommendedPaymentMethods = await shopperSession.isEligibleInPaypalNetwork({
-      email: "email@test.com",
-      phone: {
-        countryCode: "1",
-        nationalNumber: "2345678901",
-      },
-    });
+    const recommendedPaymentMethods =
+      await shopperSession.isEligibleInPayPalNetwork({
+        email: "email@test.com",
+        phone: {
+          countryCode: "1",
+          nationalNumber: "2345678901",
+        },
+      });
 
     expect.assertions(1);
     expect(recommendedPaymentMethods).toEqual(true);
   });
 
-  test("should return isEligibleInPaypalNetwork false if all payment methods are false", async () => {
+  test("should return isEligibleInPayPalNetwork false if all payment methods are false", async () => {
     const shopperSession = createShopperSession({
       request: () =>
         Promise.resolve({
@@ -122,19 +122,20 @@ describe("shopper insights component - isEligibleInPaypalNetwork()", () => {
         }),
     });
 
-    const recommendedPaymentMethods = await shopperSession.isEligibleInPaypalNetwork({
-      email: "email@test.com",
-      phone: {
-        countryCode: "1",
-        nationalNumber: "2345678901",
-      },
-    });
+    const recommendedPaymentMethods =
+      await shopperSession.isEligibleInPayPalNetwork({
+        email: "email@test.com",
+        phone: {
+          countryCode: "1",
+          nationalNumber: "2345678901",
+        },
+      });
 
     expect.assertions(1);
     expect(recommendedPaymentMethods).toEqual(false);
   });
 
-  test("should return isEligibleInPaypalNetwork false if no eligible payment methods", async () => {
+  test("should return isEligibleInPayPalNetwork false if no eligible payment methods", async () => {
     const shopperSession = createShopperSession({
       request: () =>
         Promise.resolve({
@@ -142,13 +143,14 @@ describe("shopper insights component - isEligibleInPaypalNetwork()", () => {
         }),
     });
 
-    const recommendedPaymentMethods = await shopperSession.isEligibleInPaypalNetwork({
-      email: "email@test.com",
-      phone: {
-        countryCode: "1",
-        nationalNumber: "2345678901",
-      },
-    });
+    const recommendedPaymentMethods =
+      await shopperSession.isEligibleInPayPalNetwork({
+        email: "email@test.com",
+        phone: {
+          countryCode: "1",
+          nationalNumber: "2345678901",
+        },
+      });
 
     expect.assertions(1);
     expect(recommendedPaymentMethods).toEqual(false);
@@ -424,8 +426,6 @@ describe("shopper insights component - validateSdkConfig()", () => {
           ...defaultSdkConfig,
           sdkToken: "",
           pageType: "",
-          userIDToken: "",
-          clientToken: "",
         },
       })
     ).toThrowError(
@@ -440,28 +440,10 @@ describe("shopper insights component - validateSdkConfig()", () => {
           ...defaultSdkConfig,
           sdkToken: "sdk-token",
           pageType: "",
-          userIDToken: "",
-          clientToken: "",
         },
       })
     ).toThrowError(
       "script data attribute page-type is required but was not passed"
-    );
-  });
-
-  test("should throw if ID token is passed", () => {
-    expect(() =>
-      createShopperSession({
-        sdkConfig: {
-          ...defaultSdkConfig,
-          sdkToken: "sdk-token",
-          pageType: "product-listing",
-          userIDToken: "id-token",
-          clientToken: "",
-        },
-      })
-    ).toThrowError(
-      "use script data attribute sdk-client-token instead of user-id-token"
     );
   });
 });